Update water_desalination_environment_v2.py
Added the possibility of changing the dense and sparse rewards: r1 is the reward if the agent doesn't finish on time (originally -1); r2 is the reward if finished on time (originally +1); r3 is the dense reward if the agent uses fossil fuel (originally -0.005); and r4 is the dense reward if the agent uses the photovoltaic cell (originally 0). By default, the instance of the environment object is created with the original values.